+++ ethers-v5/src/constants.chains.ts
@@ -1,0 +1,18 @@
+import type { Networkish } from "@ethersproject/networks";
+
+/**
+ * harcoded mapping of Direct.dev network ID --> Ethers Networkish
+ */
+export const networks: Record<SupportedNetworkId, Networkish> = {
+  ethereum: "mainnet",
+  "ethereum-holesky": "holesky",
+  "ethereum-sepolia": "sepolia",
+  sonic: {
+    name: "sonic",
+    chainId: 146,
+  },
+  "sonic-testnet": {
+    name: "sonic-testnet",
+    chainId: 64_165,
+  },
+};